Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I reverted the original commit due to angry emails of users getting gigantic roaming bills. I am a telephony noob, and have generally avoided touching any of the telephony code in Android for fear of exactly this.

The description given in this option makes absolutely no sense to me. I don't understand what it's for, other than the fact that a bunch of people in Canada will get huge phone bills if they check it.

Patch-set: 1