Update patch set 1
Patch Set 1:

Perhaps a warning popup like the "Dragons ahead" dialog in the Performance settings would be a good idea to warn users about the potential impact of the patch. Kind of like the stock Roaming Data checkbox. After all, it's a bit hard to explain what this settings does in a single short line.

Patch-set: 1