Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

This is an almost necessary feature.  In cold or hot weather, the touch screen doesn't respond correctly, either because the screen electronics fail and/or because someone's fingers are gloved, dried up, filled with chemical of some sort, or too wet, or very often when the phone is SO bogged down that it can't remember a full swipe while you're doing it or it detects the swipe too slowly, and the caller hangs up or goes to voicemail, and you loose business and/or a social call.

Trackball to Answer bypasses all of those problems, and saves a lot of calls.  PLEASE APPROVE FOR INCLUSION.  This is a demand and an approval of a vastly important feature.

Patch-set: 1