phone: Add option for setting device phone number Author:cyanogen
phone: Fix case where MSISDN alpha tag is not set

The new "My phone number" option was not being saved to the SIM because
the alpha tag was not previously set. This change externalizes the
preference code and defaults the alpha tag to "Voice Line 1".
Author:cyanogen

Change input type to phone number for 'My phone number' dialog in
GSM additional options.
Author:mtwebster

Change-Id: If2ba0134c737060a2efff411a589b1a132e2686c