Update patch set 1
Patch Set 1:

Then I propose you take one more look at this proposal. This adds an option to the phone settings to disable the notifications entirely (or reenable them). The reason I proposed this is because getting that false alarm over and over on sprint/cdma-like phones gets rather annoying. It also defaults to the current behavior, so nothing will be lost.

Patch-set: 1