Update patch set 1
Patch Set 1:

I apologize to Steve, Evan, and any future reviewers for splitting this patch. I was having trouble with git/repo and wound up with two separate change submissions, so I thought it was just as easy to use this updated commit as opposed to the old one.

Patch-set: 1