Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Why the new namespace? Also, coding style calls for 4 spaces instead of tabs. Why not just store the set value into a system property instead of on shutdown (which is not guaranteed to happen).

I'd really like to find the real reason why this stuff isn't being saved to the SIM.

Patch-set: 1