Update patch set 1
Patch Set 1:

It happens on Vision sometimes, but only when overclocking. It's the same issue as the screen-on animation- racy code in PowerManagerService.

I've not seen this happen on N1 though, the issue I am seeing reported on N1 is the a1026 driver failing to start because of the i2c bus being busy (race condition?). When this happens for you guys, have you checked dmesg for these i2c retries?

Patch-set: 1