Update patch set 1
Patch Set 1:

This error often appears basically horizontal orientation 998 and the phone activated and maintained in a horizontal position to reproduce the bug you press the power button to lock the terminal and if it receives an incoming call buttons do not appear to answer the call, if not return fails to unlock the terminal and re-locking the power button, between the second and fourth ring appear fails to call controls.
It is not the same as the A1026 failure is different
Also no problem by removing the comments of the source.
Sorry for my poor english

Patch-set: 1