Update patch set 1
Patch Set 1:

If you look at hideIncomingCallWidget method in InCallTouchUi.java, you can see that it runs an alpha animation that slide the alpha from 1 to 0. When this method ends the alpha value for IncomingCallWidget (Sliding or Rotatory) is 0, and Visibility property is GONE. 
Well, you can see at showIncomingCallWidget that Visibility property is set to VISIBLE but the alpha value is still 0.

To fix it, the showIncomingCallWidget method would be set alpha value to 1 to make IncomingCallWidget visible.

You have to remember that a View can have Visibility property set to VISIBLE but if the alpha value is 0 the view never will be showed.

The bug is there.

Of course, feel free to erase the names from the source, we write it to have control about temporary changes.

Thanks in advance.

Patch-set: 1