Update patch set 1
Patch Set 1:

Nobody would search it In the phone settings. and then toggle hide arrows in cm settings.

I do agree a bit, that that place is not 100% perfect, but hiding it outside of cm settings would be worth.

I personally think, all modified functionality should be in cm options - users search it there and also migration from 61 to 70.missed a lot of stuff that was not under cm options (like silent state) and some didn't make it to cm 7 yet (like pattern unlock message)

Either we slightly modify the lock screen menu description or add a new point.- or keep as is

Patch-set: 1