Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Phone has the code to switch orientation built in, but there has been a design decision to force us to remain in portrait.  Landscape would be useful for speakerphone, bluetooth headsets, or in a car mount.  Everything works and looks as expected in landscape.  The proximity sensor will keep the screen from rotating when it is by your ear.

Patch-set: 1