Update patch set 2
Patch Set 2:

This is useful for speakerphone operation, bluetooth headset, when in a car dock, when you don't want to read sideways.  For devices with a kickstand (supersonic) this allows operation when in landscape and reduces unnecessary orientation changes.  If you get a call when HDMI mirroring you cannot answer it without going to the phone as the mouse orientation switches with the app orientation.  All in all, I don't see any reason to not include it as it defaults to off.

Patch-set: 2