Update patch set 2
Patch Set 2: No score

wildcards only work for starting, blocking whole carrier.
for example my number is 01767xxxxxx and i live in germany with +49 code.
it only blocks if i put 0176.* but not if i put +49176.*
i tried to put like *.176.* but it dont work like that.
and when i put *.XXXXX for blocking, getting com.android.phone FC.

Patch-set: 2