add 1. opt. Incoming call in background; 2. ringtone_query_wait_time property
1. Option to display only notification/play ringtone when incoming call
arrives if screen is on. Useful when working on something, so the current
top app is not put in background automatically. User can save important work
before picking up the call without risking the app to be killed while in
background during the call.

2. Allow to override the default max query time before giving up and falling
back to the default ringtone, via ro.ringtone_query_wait_time property.
On some devices in some situations, 500ms is not enough. It can be used
to help the issue when customized, contact specific ringtone is not being
played too often on certain devices.

Change-Id: I34825afb64e64502e07ad10bcc62e6f089f1a5fe