Update patch set 1
Patch Set 1:

Yup. I tried it out and the per-contact ringtones and they worked (CDMA phone) and also tried the blacklist (which was reported to me as not working), and it does.  I'd certainly like to see your patch committed.

I believe the race is in Ringer.java (hence I invalidated the ringtone object).  Once the ringtone object is populated it's set.  If the ringtone URI is changed, you've still got the old ringtone object.  One path will set the default systemwide ringtone, the other will set the custom one.

Patch-set: 1