Update patch set 1
Patch Set 1:

I'm not familiar with the PBAP code and where it comes into play. Is it just to show the contact name when there is an incoming call? If so, I don't think we would want to filter things out. Let me know if I should remove the extra space or major changes are needed as you implied here.

Patch-set: 1