Update patch set 3
Patch Set 3:

This whole class was written by CM so you can't blame Google ;)

What would be the advantage of TextUtils.equals rather than just the built-in java method? IMO TextUtils' version is just for those using CharSequence which has no .equals method. We're using Java's String which has the .equals method and IMO, it'd be better and more readable to use that.

Patch-set: 3