Update patch set 3
Patch Set 3:

Danny,

On Milestone, the summary used by Motorola is "Optimize in-call audio profile".

The wording used in this patch suggests that the user will hear a difference when he change the setting, which is misleading. Even in the case it is formulated this way by Motorola on Defy, it's still wrong :).

Patch-set: 3