Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I was not sure if it is common to code review yourself, but I saw someone else do it so here goes.

I guess there are some other ways I could have done this. I could have tried to check if the file exists using the File class, but I would not be sure if the URI is a file, although generally it is a good assumption. I could also have altered RingtoneManager to have automatically selected a default if the original fails (as it says it should do in the API documentation) but I have no way of knowing if we are requesting a full ringtone or a notification tone as far as I can tell.

This seemed to be the simplest and most effective way to do it. If it cannot find a Ringtone for the URI, try the default.

Possibly a future improvement that may hit more skepticism is if a ringtone is not available, turn on vibrate on the phone. I figured some people may object to that, so I decided not not even try to put that in, since this is the core issue.

Patch-set: 1