Change proximity behaviour for devices w/ fixed keyboard
The actual behaviour of Phone, force screen to stay On when the Keyboard is
open, while correct for slidding keyboard, is not correct in the case of a
device with a fixed open keyboard, like the HTC Chacha.

This patch to PhoneApp.java make it behave differently if the device has a
fixed/always open physical keyboard, it will not check if keyboard is open or not.

Set config_device_has_fixed_keyboard to true in the overlay of the device.

Change-Id: Ic7ad89ac35271136a68b6ef8901028418c7c43e5