Change proximity behaviour for devices w/ fixed keyboard
The actual behaviour of Phone, force screen to stay On when the Keyboard is
open, while correct for slidding keyboard, is not correct in the case of a
device with a fixed open keyboard, like the HTC Chacha.

This patch to PhoneApp.java make it behave diferently if the device has a
fixed/always open physical keyboard. If ro.product.has_fixed_keyboard is set
to 1, it will not check if keyboard is open or not.

Change-Id: Ia3b635dca889508404d21a29c014cc5b0413b8b0

Signed-off-by: Alexandre Dumont <adumont@gmail.com>