Update patch set 1
Patch Set 1:

In src/com/android/phone/PhoneApp.java current code shows:

                 boolean screenOnImmediately = (isHeadsetPlugged()
                             || PhoneUtils.isSpeakerOn(this)
                             || ((mBtHandsfree != null) && mBtHandsfree.isAudioOn())
                             || mIsHardKeyboardOpen);

mIsHardKeyboardOpen is true (which is correct), so I get screenOnImmediately = true, and so it will not check proximity sensor, which is not the expected behaviour.

    public void onConfigurationChanged(Configuration newConfig) {
        if (newConfig.hardKeyboardHidden == Configuration.HARDKEYBOARDHIDDEN_NO) {
            mIsHardKeyboardOpen = true;
        } else {
            mIsHardKeyboardOpen = false;
        }

mIsHardKeyboardOpen is true because newConfig.hardKeyboardHidden == Configuration.HARDKEYBOARDHIDDEN_NO
Actually IMHO that is correct, the keyboard is open / the hardware keyboard isn't hidden.
I don't see why I would want to change that.

Actually from my understanding, from what you quote in PhoneWindowManager.java#L957, 

    private int determineHiddenState(boolean lidOpen,
            int mode, int hiddenValue, int visibleValue) {
        switch (mode) {
            case 1:
                return lidOpen ? visibleValue : hiddenValue;
            case 2:
                return lidOpen ? hiddenValue : visibleValue;
        }
        return visibleValue;
    }

I believe in my case of a device with always open kbd, it's correct to use config_lidKeyboardAccessibility = 0 and don't care if there is a lid, or not or if the kernel reports a lid or not correctly, because it wouldn't make much sense anyway. The device doesn't have a lid. In determineHiddenState(), mode=config_lidKeyboardAccessibility=0 which leads to hardKeyboardHidden=visibleValue=Configuration.HARDKEYBOARDHIDDEN_NO.

What I pretent to change is the way PhoneApp checks if it will stay screenOn (via screenOnImmediately) or check the proximity sensor, because for the moment one rule is based on if the keyboard is open or not. There is no posibility to "bypass" this rule. This is what I pretend to add with this patch.

On the other hand, I fail to see how changing lid state, and/or config_lidKeyboardAccessibility will affect that (unless it makes newConfig.hardKeyboardHidden NOT igual Configuration.HARDKEYBOARDHIDDEN_NO , which would probably lead to issues elsewhere in other apps I can guess).

Patch-set: 1