Update patch set 1
Patch Set 1:

You are fully correct in your analysis, we just draw different conclusions. AFAICT, the main question is: if KEYBOARDHIDDEN_NO is set, do apps expect
- any kind of hardware keyboard being present or
- a slidable hardware keyboard which happens to be open?

Given that Android devices with a permanently open keyboard are fairly new, I would think it's the latter. You think it's the former. We both don't know what the right answer is - or if there is a right answer at all.

Having said that, I'm ok-ish with overlaying that behaviour in config.xml. Did you have a look how the problem is solved for the LG C660 (similar device, just portrait display, which shouldn't matter here)?

Patch-set: 1