Update patch set 12
Patch Set 12:

Rick,
When you have a time to test this again, please try to send me a logcat. Because there's some string being logged that can help us to view the point of the code (for both PhoneToggler.java and Settings.java) is being used by your phone. If the logcat is too big, send to my email or to pastebin...
Tks.

Patch-set: 12