Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

fiouzy,

There's already a working in progress to port this.

The changeset below it's complete with all the strings needed and commit history, and with some fixes needed for LTE phones being tested. If you want to collaborate, you can test and review:

http://review.cyanogenmod.com/#change,12820

Regards

Patch-set: 1