Update patch set 6
Patch Set 6:

Danny, we need this on a per-device basis because this is a crippling albeit rare issue on the Sprint network.  This option defaults devices built with this option to mimic Sprint's Android ROM behavior of ignoring Message Wait Indicators.  Without this fix, you see hundreds of users replacing Phone.apk with a random build from the forums where someone had done a smali edit.

It cannot be a hard-coded behavior however because there are less common instances where users of this ROM actually want MWI's, usually in cases where they are using a Sprint phone on an overseas carrier.

Justin has been busy with school, but now he has time to finish this patch according to all of our comments.

Patch-set: 6