Update patch set 1
Patch Set 1: Restored

Fix doesn't seem to be permanent. I'm not sure why, after the first call the custom ringtone seems to start playing.


Not sure if it is a caching issue for the sounds, or what. Some seem to play first time after a reboot, others... not so much seems really intermittent.

But they do play so.

Patch-set: 1