Update patch set 2
Patch Set 2:

I tested for that specific scenario, setting an LTE mode, rebooting, even wiping, and it didn't reset the network mode when entering that screen or opening the preference.

Checking the MobileNetworkSettings::onPreferenceChange method, there is this ominous comment "NOTE onPreferenceChange seems to be called even if there is no change".  I'm interpreting that (and I don't otherwise see how) as the method may be called when the network mode preference is opened (and thus, the values are set correctly) but the user cancels or backs out.  I'm not seeing how the value may be reset otherwise, except when it's not recognized as a valid option at all by handleGetPreferredNetworkTypeResponse().

If you wouldn't mind giving it a try when you have a chance, that would be great.  It seems to unnecessary to include the LTE mode settings on devices that don't advertise LTE support.  But of course, if it does actually break devices or if you would just prefer to keep that list comprehensive, I can change it.

Patch-set: 2