Update patch set 5
Patch Set 5: Looks good to me, but someone else must approve

I'm one of the authors of the ICS-Forward Ported Code. I started working on the forward port to jb and am glad someone else already took care of it. 

Vibrate 45 works, Vibrate on Answer and Vibrate End works. I haven't tested Vibrate on Call Waiting, but I'm assuming it works, as I did test it extensively on ICS and this is the same code :)

Patch-set: 5