Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

The 'Incoming call in background' option was meant as a possibility to protect work being done on phone, as the app one was working in could get killed by oom killer when suddenly moved to background by incoming call. So that option allows to display only incoming call notification, but not fire the fullscreen intent for in-call ui when the screen is on at the time of incoming call.

But when the phone is in dock, the screen is usually on all the time, regardless whether one is doing something on the phone or not. The fact that the phone is docked rather suggests that the phone is not being directly used by user.

Therefore I think that this addition of docked state check is only logical.

My personal reason for this change:
When running navigation app like e.g. Sygic in full screen mode while driving (phone in car dock), notifications are not visible at all. I want to have the 'Incoming call in background' enabled, but I need to see who's calling me while driving.

Patch-set: 1