PhoneApp: Option to not lock automatically when a call ends (2/2)
Depends on http://review.cyanogenmod.com/23251

When a call ends while the screen is off the phone is automatically locked. This commit adds an option to disable this behavior.

Navigate to Phone->Settings and uncheck the option Lock when call ends under the advanced settings section.

Commit includes option in english and translation to portuguese.

Change-Id: Ic0894bc5095fc8263ca350fb6b7ffcad494a4050