Update patch set 4
Patch Set 4:

My personal opinion is that, while this patch is useful, the settings should not be able to be enabled if there is a passcode lock on the device.

Eg. If the phone was stolen, the thief would be able to make a call to then phone and then would have access to all the phone functions. This would not happen with the current setup.

Patch-set: 4