Update patch set 4
Patch Set 4: No score

@Lalit : This option do not disable to screen lock. If the phone is locked and someone calls you it will still be locked. If you press home in this situation, the screen lock will be brought up as it should.

The AOSP default behavior that is changed by this option applies only when the phone is already unlocked. If you have the phone unlocked and have a call disconnected with the screen off because of the proximity sensor ( you have the phone in your ear ) then AOSP default behavior is to lock the phone immediately. This option allows you to change this behavior and keep the phone unlocked when the call is disconnected in this scenario.

If someone steals your phone and make a call to it it will still be locked when the call ends.

I there some situation I'm missing ?

Regards

Patch-set: 4