Update patch set 5
Patch Set 5:

My opinion on this patch. If I have a passcode on the phone, have it set to not automatically lock when a phone call ends & make a call, then it should not go into a locked state and remain in the unlocked state, but if I receive a call instead, then the phone should go back to a locked state after the call is finished.

Patch-set: 5