Fix call end vibration.
As on call end both the vibration was started and the vibration thread
was killed, there was a race condition where call end vibration
sometimes worked and sometimes it didn't.
As the custom vibration thread handling was unnecessary anyway, remove
it and instead create a pattern for the vibrator (which internally will
use a thread again).

Change-Id: I5574737b94104e990ee8b08349555571217902c4