Update patch set 4
Patch Set 4: Looks good to me, but someone else must approve

I have moved the majority of the CM strings into a separate category at the bottom, save for the ones that are modifications to the original AOSP strings, and the ones that fit into AOSP categories. No related build errors, and I believe this is good to go.

Patch-set: 4