Update patch set 1
Patch Set 1:

In the future, you don't have to abandon the patch if the change-id doesn't end up at the bottom. Just amend the commit locally and change the last line to match what is on gerrit. Sometimes change-id's don't show up, you can get it in the Changes info in the top left corner, above Owner.

Patch-set: 1