Update patch set 17
Patch Set 17:

@Dani:

I have experienced strange things. My test cases:

1. 
1.1 Turn your phone face down. 
1.2 Call your phone
The result: The phone started to ring, and few (2-3) seconds later the ringing ended. The result is the same with the dismiss call option. When your phone is in your pocket, the ringing will be muted accidentally.

2. 
2.1 Call the phone (the ringing started)
2.2 Turn the phone face down (the ringing ended)
2.3 Turn the phone face up
2.4 Hang up the phone
2.5 Call the phone again immediatly
The result: The ringing started and ended a second later. (The phone is turned to face up)

Patch-set: 17