Update patch set 17
Patch Set 17: I would prefer that you didn't submit this

The weird thing is that this never happens with DeskClock. When it's turned down, it doesn't do anything. The sensor code is exactly the same here so I'm not sure yet what the real problem is. Thank you for reporting it, I was able to reproduce, too.

Patch-set: 17