Update patch set 1
Patch Set 1:

We may also consider ripping out the 'feature' of going to the call log after a call entirely. Can anyone of you guys think of a valid use case for this? AOSP has written a _lot_ of code for it, but I wonder what the point of it is.

Patch-set: 1