Update patch set 1
Patch Set 1:

This has been bothering me as well, I always feel that after a call I should be back at whatever I was doing before. Seems odd that AOSP would disrupt a workflow like that. A toggle is probably best, but can the default be to "return to previous activity"?

Patch-set: 1