Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

For references, this was added by Google to "save network mode as much as possible" since crespo's ics radio does not save network mode. However it is a bit overboard as the network mode is saved on set anyway.

Patch-set: 1