Update patch set 4
Patch Set 4: Looks good to me, but someone else must approve

I agree with this approach - the Android design guidelines supports using the summary to indicate the current state of a setting - I have started doing it in many places in settings, Quiet Hours being one of them.

As for Quiet Hours, some of the back end framework stuff came over from CM7 but I completely rewrote/redesigned the UI components of it to be more compliant with the guidelines and use modern coding practices.

Patch-set: 4