Update patch set 18
Patch Set 18:

I tested on my i9300 and if my screen is locked then the calls are droped. If screen is unlocked the calls are only muted. Is it okay? And if the call is matching the regexed blocklist is it need to show remove block button in the notification?

Patch-set: 18