Update patch set 18
Patch Set 18:

Tamás, it should hang up when a phone is blocked regardless of what the state of the screen is.  It uses the aosp phone internal hangup function. 

Good catch on the regex.  Hmm, not sure how it should be handled by the notification, if at all. And it it doesn't, we would have an inconsistent UI.  Maybe we should just drop the whole notification action idea?

Patch-set: 18