Update patch set 18
Patch Set 18:

David I tested several times but if the phone was unlocked (and not just the screen was turned on with the power button) the call was muted only. I was hearing the ringback on the other phone. If the phone was screen locked, then the call was dropped imediatly.

The notification is OK but I think the 'Remove block' button is not needed, if you couldn't resolve the UI inconsistent. And one more question in relation to the regex. Do you plan to be add exception handling or WhiteList too? Or how to add one or more numbers to the 123.* list which call's you don't want to reject?

Patch-set: 18