Update patch set 19
Patch Set 19:

David:I tested the PS19. Now it's only offered the "Remove Block" button, when the whole number is matching. Thanks for it. But the mute/drop situation is still there. Suggestions: If the number is in the contacts we could display the name instead of the number.

Patch-set: 19