Update patch set 19
Patch Set 19:

David: I think I was not crystal clear. On the called phone (which is running with CM) everything is working perfectly. In both cases only a blacklist notification is displayed. But on the caller phone in one case the call will be terminated immediately in the other case the call will ring as if I didn't want to answer the call.

Patch-set: 19