Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

This type of change should be submitted to the AOSP gerrit instance not CM.  It needs to be fixed in the original source or we will just have to re-apply it every time a new version of Android is released by Google.

Please submit to AOSP and abandon it here.

Patch-set: 1