Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Not sure whether it's better to add the button to the landscape layout instead of the null checks, but I couldn't find a good place for it: first row is a 2-column audio mode button, second row is mute and hold, third is dialpad and add, fourth is 2-column end call. This button is probably not important enough that we need to care, especially given that the landscape layout is only used in docks.

Patch-set: 1