Update patch set 1
Patch Set 1:

The reason why it's confusing is that the "Add to blacklist" option is shown in various places, but it doesn't actually do anything unless you dig down into the settings and turn it on.

Gotta think like a user here, it's not very discoverable. If I click "add to blacklist" and I get a call and it's not blacklisted I then think "Oh fuck this broken shit", not "Oh there's probably a hidden setting for that!" :)

Patch-set: 1