Update patch set 1
Patch Set 1: Code-Review+1

As a user of this feature, I actually just found out yesterday that Steve is almost exactly right down to the words used--I have a long list of 800 numbers that I blacklist, but for the longest time I've been waiting and searching for why the blacklist feature didn't seem to work.

Now I know why. I didn't turn it on.

Patch-set: 1
Label: Code-Review=+1