Update patch set 1
Patch Set 1:

Ricardo,

it's a setting to make sure UI attached to the blacklist feature (the 'add to blacklist' button, actually) is hidden if the blacklist feature is not desired to be used in order to not accidentially blacklist callers.

Steve,

do you really want me to -2 patches if I have issues with their correctness? :( As it stands (and mentioned above), this patch is plain broken: If the user didn't touch the setting, blacklist (and its notification) will show as disabled in the settings UI, but actually is enabled. If the user touches the 'add to blacklist' button by accident (or because he wanted to try what this button does), future calls will be rejected, even though the blacklist seems to be disabled.

I'll prepare a revert.

Patch-set: 1