Update patch set 1
Patch Set 1:

And about Steve's quote:

"
Gotta think like a user here, it's not very discoverable. If I click "add to blacklist" and I get a call and it's not blacklisted I then think "Oh fuck this broken shit", not "Oh there's probably a hidden setting for that!" :)
"

this is what actually should _not_ be happening. If the setting is disabled, this button simply should not be there. If it is, it's a bug (which I don't see myself - no button visible here).

Patch-set: 1