Update patch set 1
Patch Set 1:

Go to call log, click on item, hit menu. "Add to blacklist" shows even though the blacklist is disabled. This is going to be how most users use it, and it's broken so then there are multiple issues going on here. What we want is- blacklist *on* by default (it's a very useful feature and users shouldn't have to dig for it), and no showing of those options if it's off.

Patch-set: 1