Update patch set 1
Patch Set 1:

The problem is that the option you're mentioning is in the contacts app, which doesn't have access to the phone app's shared prefs. To fix this, I think we need to make the blacklist enable state a system setting.

This doesn't change the fact that this patch is broken, though.

Patch-set: 1