Update patch set 1
Patch Set 1:

I have no deep emotional attachment to this question, but I can say that in the 2.5 years I'm using Android now I haven't had the need for blacklisting a caller a single time. I guess the usefulness of the feature depends on country and/or how widespread one's number is.

Patch-set: 1