Update patch set 1
Patch Set 1:

But if all it does is sends the stalker or telemarketer to your voicemail, the blacklist is useless as well.  You may just as well just not answer the phone and it will go to voicemail automatically after a few rings.  It defeats the purpose of the blacklist which is to leave no ability for the calling entity to know that they have contacted you or confirm that your number is a valid number by letting it ring.

there must be another way around this problem ...

Patch-set: 1