Update patch set 3
Patch Set 3:

It says that file was changed 22 days ago, but I can confirm that it did not set the default for the auto complete.  Do a fresh install from 10.1 -> 10.2 and you will see.  I feel it will cause confusion for users upgrading as they will seem to lose functionality unless the setting is on.

Patch-set: 3