Update patch set 2
Patch Set 2:

One does not simply fix upstream typos.

Seriously, and this is a good example of why: a trivial misype took you 4 patches across 4 projects to fix. Don't you think qualcomm knows this typo is here? :) It's just not worth it, and for us, as a downstream, it means that if we mess up one of these 4 projects when merging a new version, we'll spend a while tracking down the reason why.

There are some non-visual errors (typos, line breaks, whitespace, and the like) that, as annoying as they may be, are just not worth fixing... unless they're actually our fault.

Patch-set: 2
Reviewer: Ricardo Cerqueira <10@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=0