Phone: Honor the "Rejected calls as missed" option
When the type of a call is set, either for the purposes of call logging or to
display a notification for a missed call, the value of the "Rejected calls as
missed" option should be checked as well. However, in CM 10.2 this check is
missing, and thus the option has no effect. This patch reintroduces the check.

Change-Id: I01ae90d3c9335e1c92af4c9c76e7b53b7e0fa8aa