Update patch set 4
Patch Set 4:

Danny, thanks for having that quick look -- I reverted the piece of code you noted. And, sure, take your time with the review, guys, and thanks a lot for your great work on CM!

Vaibhav, I wonder if the other change should be better marked as abandoned, so it doesn't cause confusion in future?

Patch-set: 4